+++ src/components/Member/DogInfo/DogDetailInfoEdit/styles.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export { Card, Stack, Caption } from "components/Enrollment/Form/styles";
+export { Card, Stack, Caption } from "components/Enrollment/NewForm/styles";
 export { Label } from "components/common/Title/style";
 export { Text } from "components/common/Text";
 export { Content } from "components/Enrollment/styles";
